requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py310h2bbff1b_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py310haa95532_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.4=py310h9128911_0
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py310h2bbff1b_1002
bzip2=1.0.8=he774522_0
ca-certificates=2022.4.26=haa95532_0
certifi=2022.5.18.1=py310haa95532_0
cffi=1.15.0=py310h2bbff1b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py310haa95532_0
colorama=0.4.4=pyhd3eb1b0_0
confuse=1.5.0=pyhd3eb1b0_0
cryptography=37.0.1=py310h21b164f_0
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.5.1=py310hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py310haa95532_0
executing=0.8.3=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
htmlmin=0.1.12=pyhd3eb1b0_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=3.3=pyhd3eb1b0_0
imagehash=4.2.1=pyhd3eb1b0_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.9.1=py310haa95532_0
ipython=8.3.0=py310haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py310haa95532_1
jinja2=3.0.3=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h2bbff1b_0
jsonschema=4.4.0=py310haa95532_0
jupyter=1.0.0=py310haa95532_7
jupyter_client=7.2.2=py310haa95532_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_core=4.10.0=py310haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.4.2=py310hd77b12b_0
libffi=3.4.2=h604cdb4_1
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=he0120a3_1
libuv=1.40.0=he774522_0
libwebp=1.2.2=h2bbff1b_0
llvmlite=0.38.0=py310h23ce68f_0
lz4-c=1.9.3=h2bbff1b_1
markupsafe=2.1.1=py310h2bbff1b_0
matplotlib=3.5.1=py310haa95532_1
matplotlib-base=3.5.1=py310hd77b12b_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
missingno=0.4.2=pyhd3eb1b0_1
mistune=0.8.4=py310h2bbff1b_1000
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py310h2bbff1b_0
mkl_fft=1.3.1=py310ha0764ea_0
mkl_random=1.2.2=py310h4ed8f06_0
munkres=1.1.4=py_0
nbclient=0.5.13=py310haa95532_0
nbconvert=6.4.4=py310haa95532_0
nbformat=5.3.0=py310haa95532_0
nest-asyncio=1.5.5=py310haa95532_0
networkx=2.7.1=pyhd3eb1b0_0
nltk=3.7=pyhd3eb1b0_0
notebook=6.4.11=py310haa95532_0
numba=0.55.1=py310h4ed8f06_0
numexpr=2.8.1=py310hb57aa6b_1
numpy=1.21.5=py310h6d2d95c_3
numpy-base=1.21.5=py310h206c741_3
openssl=1.1.1o=h2bbff1b_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=py310hd77b12b_0
pandas-profiling=2.9.0=pyhd3eb1b0_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
phik=0.11.2=pyhd3eb1b0_1
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=py310hdc2b20a_0
pip=21.2.4=py310haa95532_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
pure_eval=0.2.2=pyhd3eb1b0_0
pycocotools=2.0.4=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py310hd77b12b_6
pyrsistent=0.18.0=py310h2bbff1b_0
pysocks=1.7.1=py310haa95532_0
python=3.10.4=hbb2ffb3_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
pytorch=1.11.0=py3.10_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pywavelets=1.3.0=py310h2bbff1b_0
pywin32=302=py310h2bbff1b_2
pywinpty=2.0.2=py310h5da7b33_0
pyyaml=6.0=py310h2bbff1b_0
pyzmq=22.3.0=py310hd77b12b_2
qt=5.9.7=vc14h73c81de_0
qtconsole=5.3.0=pyhd3eb1b0_0
qtpy=2.0.1=pyhd3eb1b0_0
regex=2022.3.15=py310h2bbff1b_0
requests=2.27.1=pyhd3eb1b0_0
scipy=1.7.3=py310h6d2d95c_0
seaborn=0.11.2=pyhd3eb1b0_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=61.2.0=py310haa95532_0
sip=4.19.13=py310hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.38.3=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
tangled-up-in-unicode=0.1.0=pyhd3eb1b0_0
tbb=2021.5.0=h59b6b97_0
terminado=0.13.1=py310haa95532_0
testpath=0.5.0=pyhd3eb1b0_0
tk=8.6.12=h2bbff1b_0
torchvision=0.12.0=py310_cu113
tornado=6.1=py310h2bbff1b_0
tqdm=4.64.0=py310haa95532_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py310haa95532_0
vc=14.2=h21ff451_1
visions=0.5.0=pyhd3eb1b0_0
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py310haa95532_1
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py310haa95532_0
win_inet_pton=1.1.0=py310haa95532_0
wincertstore=0.2=py310haa95532_2
winpty=0.4.3=4
xz=5.2.5=h8cc25b3_1
yaml=0.2.5=he774522_0
zlib=1.2.12=h8cc25b3_2
zstd=1.5.2=h19a0ad4_0

numpy~=1.21.5
matplotlib~=3.5.1
torchvision~=0.12.0
pillow~=9.0.1
tqdm~=4.64.0
nltk~=3.7
pycocotools~=2.0.4